requirements.txt
numpy>=1.16.1
pandas>=0.23.4
pyahocorasick>=1.1.12